Вынес Router в отдельный компонент
